refactor(v0.3): split attempts controller into read/write cqrs
