+++ backend/app/Http/Requests/V0_3/StartAttemptRequest.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Http\Requests\V0_3;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class StartAttemptRequest extends FormRequest
+{
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    public function rules(): array
+    {
+        return [
+            'scale_code' => ['required', 'string', 'max:64'],
+            'region' => ['nullable', 'string', 'max:32'],
+            'locale' => ['nullable', 'string', 'max:16'],
+            'anon_id' => ['nullable', 'string', 'max:64'],
+            'client_platform' => ['nullable', 'string', 'max:32'],
+            'client_version' => ['nullable', 'string', 'max:32'],
+            'channel' => ['nullable', 'string', 'max:32'],
+            'referrer' => ['nullable', 'string', 'max:255'],
+            'meta' => ['sometimes', 'array'],
+        ];
+    }
+}
